refactor: extract model provider sync logic to separate module, add unit tests
