+++ vscode/src/main.ts
@@ -57,7 +57,7 @@ import type { PlatformContext } from './extension.common'
 import { configureExternalServices } from './external-services'
 import { isRunningInsideAgent } from './jsonrpc/isRunningInsideAgent'
 import { logDebug, logError } from './log'
-import { getChatModelsFromConfiguration, syncModelProviders } from './models/utils'
+import { getChatModelsFromConfiguration, syncModelProviders } from './models/sync'
 import type { FixupTask } from './non-stop/FixupTask'
 import { CodyProExpirationNotifications } from './notifications/cody-pro-expiration'
 import { showSetupNotification } from './notifications/setup-notification'
